Move Period overlap logic to Period class itself and move tests.
